requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
beanie==1.26.0
pymongo==4.8.0
pydantic==2.7.4
pydantic-settings==2.3.4
python-ulid==2.7.0
slowapi==0.1.9
structlog==24.2.0
openpyxl==3.1.5
httpx==0.27.2
mangum==0.17.0
python-dotenv==1.0.1
pytest==8.3.2
pytest-asyncio==0.24.0
mongomock-motor==0.0.21
